requirements.txt
altair==5.5.0
altgraph==0.17.5
annotated-doc==0.0.4
annotated-types==0.7.0
ansi2html==1.9.2
anyio==4.12.1
asgiref==3.11.1
attrs==26.1.0
backports.tarfile==1.2.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
bottle==0.13.4
branca==0.8.2
cachelib==0.13.0
cachetools==5.5.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
choreographer==1.2.1
click==8.3.1
clr_loader==0.2.10
colorama==0.4.6
colorcet==3.1.0
contourpy==1.3.3
cryptography==46.0.5
customtkinter==5.2.1
cycler==0.12.1
Cython==3.2.4
darkdetect==0.8.0
dash==2.14.2
dash-core-components==2.0.0
dash-extensions==2.0.5
dash-html-components==2.0.0
dash-leaflet==1.0.15
dash-svg==0.0.12
dash-table==5.0.0
dataclass-wizard==0.39.1
defusedxml==0.7.1
distlib==0.4.0
Django==5.2.12
docxtpl==0.20.2
EditorConfig==0.17.1
et_xmlfile==2.0.0
fastapi==0.109.0
filelock==3.25.2
Flask==3.0.3
Flask-Caching==2.3.1
folium==0.20.0
fonttools==4.61.1
fpdf==1.7.2
fpdf2==2.8.5
geobuf==2.0.1
geopandas==1.1.3
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.3.2
h11==0.16.0
httptools==0.7.1
idna==3.11
importlib_metadata==7.2.1
iniconfig==2.3.0
itsdangerous==2.2.0
jaraco.context==6.1.2
jaraco.functools==4.4.0
jaraco.text==4.2.0
Jinja2==3.1.6
jsbeautifier==1.15.4
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kaleido==1.2.0
kiwisolver==1.4.9
logistro==2.0.1
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
more-itertools==10.8.0
MouseInfo==0.1.3
mysql-connector-python==9.6.0
narwhals==2.18.0
nest-asyncio==1.6.0
numpy==1.26.4
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.3
pathlib2==2.3.7
pefile==2024.8.26
pillow==10.4.0
platformdirs==4.9.4
playwright==1.58.0
plotly==5.18.0
pluggy==1.6.0
protobuf==4.25.9
proxy_tools==0.1.0
psycopg==3.3.3
psycopg-binary==3.3.3
pyarrow==23.0.1
PyAutoGUI==0.9.54
pycparser==3.0
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.9.1
pyee==13.0.1
PyGetWindow==0.0.9
Pygments==2.19.2
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2025.11
PyMsgBox==2.0.1
PyMuPDF==1.26.7
pyogrio==0.12.1
pyparsing==3.3.2
pyperclip==1.11.0
pyproj==3.7.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.18.0
PyQt6==6.10.2
PyQt6-Qt6==6.10.2
PyQt6_sip==13.11.0
PyRect==0.2.0
PyScreeze==1.0.1
PySide6==6.10.2
PySide6_Addons==6.10.2
PySide6_Essentials==6.10.2
PySocks==1.7.1
pytest==9.0.2
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
python-discovery==1.2.0
python-docx==1.1.0
python-dotenv==1.2.2
python-multipart==0.0.26
python-pptx==1.0.2
pythonnet==3.0.5
pytweening==1.2.0
pytz==2026.1.post1
pywebview==6.1
pywhatkit==5.4
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.3
qrcode==8.2
referencing==0.37.0
reportlab==4.4.7
requests==2.31.0
retrying==1.4.2
rich==13.9.4
rpds-py==0.30.0
rtree==1.4.1
selenium==4.10.0
shapely==2.1.2
shellingham==1.5.4
shiboken6==6.10.2
simplejson==3.20.2
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
SQLAlchemy==2.0.25
sqlparse==0.5.5
starlette==0.35.1
streamlit==1.31.0
tabulate==0.10.0
tenacity==8.5.0
toml==0.10.2
tornado==6.5.5
trio==0.33.0
trio-websocket==0.12.2
ttkbootstrap==1.20.0
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.27.0
validators==0.35.0
virtualenv==21.2.0
watchdog==6.0.0
watchfiles==1.1.1
webdriver-manager==4.0.2
websocket-client==1.9.0
websockets==16.0
Werkzeug==3.0.6
wikipedia==1.4.0
wsproto==1.3.2
xlsxwriter==3.2.9
xyzservices==2025.11.0
zipp==3.23.0
